### 2019-08-21 #1669 @whedon
[PRE REVIEW]: AXITOM: A Python package for reconstruction of axisymmetric tomograms acquired by a conical beam
**Submitting author:** @PolymerGuy (<a href="http://orcid.org/0000-0002-9981-4041">Sindre Nordmark Olufsen</a>)
**Repository:** <a href="https://github.com/PolymerGuy/AXITOM" target ="_blank">https://github.com/PolymerGuy/AXITOM</a>
**Version:** V0.1.2
