### 2019-08-21 83f650a
Update paper.md
Fix to affiliation in paper.md
