### 2019-08-21 #1669 @kyleniemeyer
@polymerguy this PR should fix the compilation issue: https://github.com/PolymerGuy/AXITOM/pull/2
