### 2019-08-22 #1669 @PolymerGuy
Thank you! I have now merged the PR.
