### 2019-08-23 8f26d9c
Pre-drafts on new functionality
