### 2019-08-23 213bd5c
Fixed broken reference and improved one sentence
