+++ paper.md
@@ -17,7 +17,7 @@ bibliography: paper.bib
 
 # Summary
 The ability to picture the internals of a body in a non-destructive way is vital for both diagnostics and research in a broad range of fields.
-Tomography is the imaging of sections within a body and is typically performed by imaging of the body by a penetrating wave and subsequently reconstructing the internals of the body by employing a reconstruction algorithm. There are cases, which allows for simplification of the tomographic problem, 
+Tomography is the imaging of sections within a body and is typically performed by acquiring projection of the body by using a penetrating wave, and subsequently reconstructing the sections of the body by employing a reconstruction algorithm. There are cases, which allows for simplification of the tomographic problem, 
 one of them being the presence of symmetries within the tomogram. Axis-symmetry reduces the tomographic reconstruction into the inversion of the Abel transform, which
 can be obtained from a single projection, having a fractional acquisition cost compared to typical datasets. In home-laboratory X-ray setups,
 X-rays are distributed in space as a conical beam rather than parallel rays, and such a beam geometry has to be accounted for in the reconstruction of the tomogram.
@@ -28,7 +28,7 @@ but are general purpose and does not exploit axis-symmetries. PyAbel [@Hickstein
 
 ``AXITOM`` contains a collection of high-level functions which can be used to load datasets, perform virtual experiments and perform the tomographic reconstruction.
 A Feldkamp David Kress algorithm [@Feldkamp:84] is used to reconstruct the tomograms, and the modifications made to the algorithm are
-for reduced computational cost. The implementation relies on the use of the Numpy[@Numpy] and  Scipy[@Scipy], as well as numerous packages for visualization and IO.
+for reduced computational cost. The implementation relies on the use of the Numpy [@Numpy] and Scipy [@SciPy], as well as numerous packages for visualization and IO.
 
 ``AXITOM`` was implemented for the reconstruction axis-symmetric density fields measured by X-ray absorption radiography.
 This project is a part of the ongoing research within the SFI CASA research group at NTNU and has been a key component in the pursuit of *in-situ* 
