### 2019-08-23 #1669 @PolymerGuy
I have fixed a broken reference and improved the second sentence in the proof.
