### 2019-08-25 af50d5e
Working pre-prototype
