+++ examples/verifyAxisym.py
@@ -4,11 +4,17 @@ import axitom
 from scipy.ndimage import shift
 from scipy.ndimage.filters import median_filter
 
+from scipy.ndimage import rotate
 
+def rebin(arr, new_shape):
+    """Rebin 2D array arr to shape new_shape by averaging."""
+    shape = (new_shape[0], arr.shape[0] // new_shape[0],
+             new_shape[1], arr.shape[1] // new_shape[1])
+    return arr.reshape(shape).mean(-1).mean(1)
 
 print("pipp")
 path = "/home/sindreno/InSituCT/radiograms/"
-radiogram_names = axitom.list_files_in_folder(path)[:99]
+radiogram_names = axitom.list_files_in_folder(path)[98:]
 
 config = axitom.config_from_xtekct(path + "testSamplePVC6_1fps_gain12.xtekct")
 
@@ -41,31 +47,34 @@ for filt in filtered:
         # plt.show()
 
         # Remove some edges that are in field of view
-        radiogram[:250, :] = 0.95
-        radiogram[1800:, :] = 0.95
+        radiogram[:250, :] = 0.99
+        radiogram[1800:, :] = 0.99
 
 
 
         if filt:
-            radiogram = median_filter(radiogram, size=20)
+             radiogram = median_filter(radiogram, size=20)
 
-        _, center_offset = axitom.object_center_of_rotation(radiogram, config, background_internsity=0.9)
+        _, center_offset,angle = axitom.object_center_of_rotation_ang(radiogram, config, background_internsity=0.96)
 
+
+        radiogram = rotate(radiogram, angle, order=3, reshape=False)
 
         config.update()
         print("org",center_offset)
 
         pix_shift = center_offset / config.voxel_size_y
 
-        shifted = shift(radiogram,(-pix_shift,0),order=3,cval=0.95)
+        shifted = shift(radiogram,(-pix_shift,0),order=3,cval=0.99)
 
 
 
 
-        _, center_offset = axitom.object_center_of_rotation(shifted, config, background_internsity=0.9)
+        _, center_offset,_ = axitom.object_center_of_rotation_ang(shifted, config, background_internsity=0.96)
         print("shifted",center_offset)
 
         # shifted[shifted>0.9] = 0.
+
 
 
 
@@ -73,13 +82,21 @@ for filt in filtered:
         left = shifted[:1000,:]
 
         error = (right-left[::-1,:])
+
+        bin_image = np.abs(error[:700, :]).transpose()
+        # print(bin_image.shape)
+        bin_image = rebin(bin_image,(500,175))
+        plt.imshow(np.abs(bin_image),vmin=0.0,vmax=0.01,cmap=plt.cm.plasma)
+        # # plt.imsave("/home/sindreno/artefacts/radiogram_error_80.png",  np.abs(bin_image),vmin=0.0,vmax=0.01,cmap=plt.cm.plasma)
+        plt.show()
+
         error[right>0.9] = np.nan
 
         if filt:
-            errors_filt.append(np.nanmean(error[:700, :1700]))
+            errors_filt.append(np.nanstd(error[:700, :1700]))
             offsets_filt.append(pix_shift)
         else:
-            errors_raw.append(np.nanmean(error[:700, :1700]))
+            errors_raw.append(np.nanstd(error[:700, :1700]))
             offsets_raw.append(pix_shift)
 
 
@@ -107,7 +124,7 @@ ax1.set_ylabel("Center axis offset [pix]")
 ax2 = ax1.twinx()
 ax2.plot(errors_raw, color="red")
 ax2.plot(errors_filt, "--", color="red")
-ax2.set_ylabel("Standard deviation of error [-]",color="red")
[…]
