+++ examples/verifyAxisym.py
@@ -14,7 +14,7 @@ def rebin(arr, new_shape):
 
 print("pipp")
 path = "/home/sindreno/InSituCT/radiograms/"
-radiogram_names = axitom.list_files_in_folder(path)[98:]
+radiogram_names = axitom.list_files_in_folder(path)[99:]
 
 config = axitom.config_from_xtekct(path + "testSamplePVC6_1fps_gain12.xtekct")
 
@@ -30,7 +30,7 @@ offsets_filt = []
 errors_filt = []
 
 i=0
-filtered = [False,True]
+filtered = [False]
 
 for filt in filtered:
     for radiogram_name in radiogram_names:
@@ -53,15 +53,25 @@ for filt in filtered:
 
 
         if filt:
-             radiogram = median_filter(radiogram, size=20)
+            radiogram = median_filter(radiogram, size=20)
 
-        _, center_offset,angle = axitom.object_center_of_rotation_ang(radiogram, config, background_internsity=0.96)
+        _, center_offset,angle = axitom.object_center_of_rotation_ang(radiogram[:,:1700], config, background_internsity=0.96)
 
+        print("Center offset on unrotated image: ",center_offset)
 
-        radiogram = rotate(radiogram, angle, order=3, reshape=False)
+        radiogram = rotate(radiogram, angle, order=3, reshape=False,cval=0.99)
+        config.center_of_rot_y = center_offset
 
         config.update()
-        print("org",center_offset)
+
+#        plt.imshow(radiogram.transpose())
+#        plt.show()
+
+#        tomo = axitom.fdk(radiogram,config)
+#        plt.imshow(tomo.transpose(),vmin=0.0,vmax=0.06)
+#        plt.show()
+
+
 
         pix_shift = center_offset / config.voxel_size_y
 
@@ -71,7 +81,7 @@ for filt in filtered:
 
 
         _, center_offset,_ = axitom.object_center_of_rotation_ang(shifted, config, background_internsity=0.96)
-        print("shifted",center_offset)
+        print("Center offset on rotated image: ",center_offset)
 
         # shifted[shifted>0.9] = 0.
 
@@ -86,17 +96,17 @@ for filt in filtered:
         bin_image = np.abs(error[:700, :]).transpose()
         # print(bin_image.shape)
         bin_image = rebin(bin_image,(500,175))
-        plt.imshow(np.abs(bin_image),vmin=0.0,vmax=0.01,cmap=plt.cm.plasma)
+        plt.imshow(np.abs(bin_image),vmin=0.003,vmax=0.01,cmap=plt.cm.plasma)
         # # plt.imsave("/home/sindreno/artefacts/radiogram_error_80.png",  np.abs(bin_image),vmin=0.0,vmax=0.01,cmap=plt.cm.plasma)
         plt.show()
 
         error[right>0.9] = np.nan
 
         if filt:
-            errors_filt.append(np.nanstd(error[:700, :1700]))
+            errors_filt.append(np.nanmean(error[:700, :1700]))
             offsets_filt.append(pix_shift)
         else:
-            errors_raw.append(np.nanstd(error[:700, :1700]))
+            errors_raw.append(np.nanmean(error[:700, :1700]))
             offsets_raw.append(pix_shift)
 
 
