### 2019-08-29 #1669 @katyhuff
Thank you @PolymerGuy for these suggestions and for managing the initial pdf build issues. I think your reviewer suggestions are strong, and this package has very good hygiene. So, this should be fairly straightforward. I'm looking forward to the review process and will assign reviewers now.
