### 2019-08-29 #1669 @katyhuff
* @grlee77: Gregory Lee, you have been recommended as a reviewer for this Journal of Open Source Software submission. I think you are well suited for this review. Are you able to and interested in taking on this review?
* This submission is in need of expertise in tomography as well as familiarity with the systems and languages involved (primarily python).
* Title: AXITOM: A Python package for reconstruction of axisymmetric tomograms acquired by a conical beam
* Summary: This python package provides tools for axis-symmetric cone-beam computed tomography. A Feldkamp David Kress algorithm performs the reconstruction which have been adapted such that is exploits the axis-symmetric nature of the tomogram.
* Article Proof:  https://github.com/openjournals/joss-papers/blob/joss.01669/joss.01669/10.21105.joss.01669.pdf
* Submitting author: @PolymerGuy (Sindre Nordmark Olufsen)
* Repository: https://github.com/PolymerGuy/AXITOM
* Version: 0.1.2
