### 2019-08-29 #1669 @grlee77
I just agreed to review another JOSS publication (pypulseq) earlier today, so I am not sure how quickly I would be able to complete this one. If you cannot find another reviewer, though, you can assign me and I will try.
