### 2019-08-31 #1669 @vitorsr
> I have now looked at the list of reviewers and suggest the following based on their interests. Username: @grlee77, @PingjunChen, @vitorsr

Thank you for the suggestion @PolymerGuy. However, I am/will be somewhat busy for the next week. Please carry on with the recommendations/assignments.

I'll check back Tuesday (Sep. 3) and give a definite answer if you are not able to find another reviewer.

It would help me positively if there were a time estimate to the reviewing process (it would be my first for JOSS), and how thorough it's supposed to be in terms of mathematics and code because both can get time consuming depending on the reviewing tasks involved. Reconstruction is not a simple inverse problem, even for people familiar with it. 😋
