### 2019-09-03 #1669 @katyhuff
@dgursoy Thank you for agreeing (by email) to review this submission. I appreciate your schedule constraints and we will expect your review to begin September 15th.
