### 2019-09-03 #1669 @katyhuff
Ok folks, let's move over to #1704 for the review.
