### 2019-09-03 #1704 @katyhuff
@PingjunChen and @dgursoy : thank you for your engagement in this review process! Please see the checklists in the description of this issue, above, for guidance with your review.

@PolymerGuy, I am hoping that Pingjun Chen will be able to start their review promptly, so you should soon have some comments from that review. I have requested the expert assistance of @dgursoy, as well. That review will begin when @dgursoy becomes available in mid-September. Thank you for your patience with these reviews, @PolymerGuy.
