### 2019-09-03 #1704 @PingjunChen
@PolymerGuy Good documentation and tests in this repo. There are several minor issues I believe need to be corrected.
1. There is a RuntimeWarning when running tests, "utilities.py:34: RuntimeWarning: invalid value encountered in true_divide center_of_grav_x = np.average(np.sum(xs * covered_pixels, axis=1) / np.sum(covered_pixels, axis=1)) - m / 2." Better to get rid of it.
2. There are two copies of `example_data` in the repo, which is not the optimum solution.  Better to reconstruct `tests` and `examples` to use the same `example_data`.
3. In `README/Clone the repo` part, the usage of `You can now run an example:` can have path issue if not running the example under working directory `<path_to_axitom>/examples`. `os.path.abspath` and `os.path.dirname` can be used to solve this problem as well as the above two copies of data.
4. The font size of `Reference` in the PDF file is not consistent with `Summary`. Leave a blank line before `Reference` in `paper.md`.
