### 2019-09-06 #1704 @PolymerGuy
@PingjunChen Thank you for the feedback!
The division warning is now gone and the routine itself is now a bit more robust.
The examples will now run when executed from the main folder.

I have merged all changes into the  main branch
