### 2019-09-06 #1704 @PingjunChen
@PolymerGuy Good improvements!
@katyhuff  This python package has useful axis-symmetric tomograms reconstruction functionality, as well as detailed documentation, tests, and examples.  I recommend accepting for publication.
