### 2019-09-25 #1704 @dgursoy
@PolymerGuy great package! Thanks for working on this. I have few (most are optional) minor points:

- I had the same Runtime error that is mentioned above when installing from PyPI and running tests. I think this is because the changes are in the brach and the PyPI version hasn't been updated.
-  I had a hard time in finding the example config and data files. Also their names are different in docs and in the folder structure. It'd be nice to provide a better description of their location and names.
- In the docs, you massage the data before processing (cropping and filtering). I think it would be better if you save the cropped and filtered image saved as an example file. This should increase the clarity of your documentation.
- There is a typo in docs: "centre of rotation" should be "center of rotation".
- Config or setting object needs some documentation for others to appropriately set their parameters. These could be in the Class definition in ```config.py```. Few important ones are the units and coordinate definitions.
- I would add a short description of the module on top to guide future developers.
- Maybe ```parse``` module can be merged into the ```utility``` module.
- Please check each of the function docstrings. There are some copy+paste typos, especially in axes definitions. Some are also missing the docs (e.g. ```fdk_axisym```).
- You use ```projections`` and ```radiographs``` interchangeably. I believe both refers to the same thing, so I'd stick with one for clarity across the package.
- I'd use ```object``` or ```volume``` instead of ```data``` in ```forward_project```. Also in the same function use ```ratios``` instead of ```Ratios```?
- I'd automatically convert the input arrays to numpy arrays and give a warning if needed: ```backprojection.py", line 127, in fdk```. because this is in principle not an error, right?
- Some information about the test data can be provided. What is this image?
