### 2019-09-26 #1704 @PolymerGuy
@dgursoy Thank you for the feedback! I will implement all the improvements suggested above and i will update the PyPi package as soon as possible. I hope to have all changes done within the next couple of days.
