### 2019-09-26 1515d85
Runs correctly with anisotropic grids. Ready for cleanup
