+++ axitom/filtering.py
@@ -1,6 +1,6 @@
 import scipy.signal as sig
 import numpy as np
-
+import matplotlib.pyplot as plt
 
 def ramp_kernel_real(cutoff, length):
     """Ramp filter kernel in real space defined by the cut-off frequency and the spatial dimension
@@ -40,11 +40,14 @@ def add_weights(projections, settings):
 
 
         """
-    uu, vv = np.meshgrid(settings.detector_us, settings.detector_vs)
+    uu, vv = np.meshgrid(settings.detector_vs, settings.detector_us)
 
     weights = settings.source_to_detector_dist / np.sqrt(
         settings.source_to_detector_dist ** 2. + uu ** 2. + vv ** 2.)
 
+
+    plt.imshow(weights)
+    plt.show()
     return projections * weights[:, :, np.newaxis]
 
 
