+++ examples/comparison_to_Nikon.py
@@ -20,8 +20,8 @@ For easier comparison, a small normalization routine is used on the resulting to
 
 
 def normalize_grey_scales(image):
-    reference_grey_scale = np.average(image[250:500, 0:250])
-    background_grey_scale = np.average(image[870:1020, 280:375])
+    reference_grey_scale = np.average(image[0:250, 0:250])
+    background_grey_scale = np.average(image[870-250:1020-250, 280:375])
 
     return (image - background_grey_scale) / (reference_grey_scale - background_grey_scale)
 
@@ -31,13 +31,28 @@ def reconstruct_tomogram():
 
     radiogram = axitom.read_image(join(path_to_data, "radiogram.tif"), flat_corrected=True)
     # Remove some edges that are in field of view
-    radiogram[:250, :] = 0.95
-    radiogram[1800:, :] = 0.95
+    # radiogram[:250, :] = 0.95
+    # radiogram[1800:, :] = 0.95
 
-    radiogram = median_filter(radiogram, size=20)
+    radiogram = radiogram[250:1750,:]
+
+    plt.imshow(radiogram)
+    plt.show()
+
+    radiogram = median_filter(radiogram, size=21)
+
+    config.n_pixels_u = 1500
+    config.detector_size_u = config.detector_size_u * (1500.0/2000.)
+    config.n_voxels_x = 1500
+    config.n_voxels_y = 1500
+    config.object_size_x = config.object_size_x * (1500./2000.)
+    config.object_size_y = config.object_size_y * (1500./2000.)
+    config.update()
+
 
     _, center_offset = axitom.object_center_of_rotation(radiogram, config, background_internsity=0.9)
     config.center_of_rot_y = center_offset
+
     config.update()
 
     tomogram = axitom.fdk(radiogram, config)
