### 2019-09-26 4d93e39
Now running on an arbitrary grid
