### 2019-09-26 23a72fa
Removed some debug code
