### 2019-09-27 1a49be4
Moved range of normalization
