### 2019-09-27 01b6b4d
Removing redundant settings
