### 2019-09-27 e25c3b9
Better config
