+++ axitom/config.py
@@ -4,66 +4,24 @@ from .parse import parse_xtekct_file
 
 class Config(object):
 
-    def __init__(self):
+    def __init__(self, n_pixels_u, n_pixels_v, detector_size_u, detector_size_v, source_to_detector_dist,
+                 source_to_object_dist, angular_inc=1, center_of_rot=0, **kwargs):
         """ Configuration object which contains all settings neccessary for the forward projection
             and tomographic reconstruction using the axitom algorithm.
 
         """
 
+        self.n_pixels_u = n_pixels_u
+        self.n_pixels_v = n_pixels_v
 
+        self.detector_size_u = detector_size_u
+        self.detector_size_v = detector_size_v
+        self.source_to_detector_dist = source_to_detector_dist
+        self.source_to_object_dist = source_to_object_dist
+        self.angular_inc = angular_inc
 
-        self.n_pixels_u = 2000
-        self.n_pixels_v = 2000
+        self.center_of_rot_u = center_of_rot  # Should be in pixels!
 
-        self.detector_size_u = 400.
-        self.detector_size_v = 400.
-        self.source_to_detector_dist = 1127.464
-        self.source_to_object_dist = 76.7511959075928
-        self.angular_inc = 1.
-
-        self.pixel_offset_u = 0
-        self.pixel_offset_v = 0
-
-        self.center_of_rot_y = 0.0
-
-
-
-
-
-
-
-
-
-        self.projection_angs = np.arange(0., 360, self.angular_inc)
-        self.n_projections = len(self.projection_angs)
-
-
-
-        self.object_size_x = self.detector_size_u * self.source_to_object_dist / self.source_to_detector_dist
-        self.object_size_y = self.detector_size_u * self.source_to_object_dist / self.source_to_detector_dist
-        self.object_size_z = self.detector_size_v * self.source_to_object_dist / self.source_to_detector_dist
-
-
-        self.voxel_size_x = self.object_size_x / self.n_pixels_u
-        self.voxel_size_y = self.object_size_y / self.n_pixels_u
-        self.voxel_size_z = self.object_size_z / self.n_pixels_v
-
-        self.pixel_size_u = self.detector_size_u / self.n_pixels_u
-        self.pixel_size_v = self.detector_size_v / self.n_pixels_v
-
-        self.object_ys = (np.linspace(0,self.n_pixels_u,self.n_pixels_u, dtype=np.float64) - self.n_pixels_u / 2.) * self.voxel_size_y
-        self.object_xs = (np.linspace(0,self.n_pixels_u,self.n_pixels_u, dtype=np.float64) - self.n_pixels_u / 2.) * self.voxel_size_x
-        self.object_zs = (np.linspace(0,self.n_pixels_v,self.n_pixels_v, dtype=np.float64) - self.n_pixels_v / 2.) * self.voxel_size_z
-
-        print(self.object_xs[:])
-
-
-        self.detector_us = (np.arange(self.n_pixels_u,
-                                      dtype=np.float64) - self.n_pixels_u / 2.) * self.pixel_size_u + self.pixel_offset_u * self.pixel_size_u
-        self.detector_vs = (np.arange(self.n_pixels_v,
-                                      dtype=np.float64) - self.n_pixels_v / 2.) * self.pixel_size_v + self.pixel_offset_v * self.pixel_size_v
-
-    def update(self):
         self.projection_angs = np.arange(0., 360, self.angular_inc)
         self.n_projections = len(self.projection_angs)
 
@@ -78,14 +36,27 @@ class Config(object):
         self.pixel_size_u = self.detector_size_u / self.n_pixels_u
[…]
