### 2019-09-27 0734d60
Added docstring
