### 2019-09-27 2a12af7
Updated examples
