+++ examples/comparison_to_Nikon.py
@@ -29,16 +29,14 @@ def normalize_grey_scales(image):
 def reconstruct_tomogram():
     config = axitom.config_from_xtekct(join(path_to_data, "radiogram.xtekct"))
 
-    radiogram = axitom.read_image(join(path_to_data, "radiogram.tif"), flat_corrected=True)
+    projection = axitom.read_image(join(path_to_data, "radiogram.tif"), flat_corrected=True)
 
-    radiogram = median_filter(radiogram, size=41)
+    projection = median_filter(projection, size=41)
 
-#    config = config.with_param(n_pixels_u=1500, detector_size_u=config.detector_size_u * (1500.0 / 2000.))
-
-    _, center_offset = axitom.object_center_of_rotation(radiogram, config, background_internsity=0.9)
+    _, center_offset = axitom.object_center_of_rotation(projection, config, background_internsity=0.9)
     config = config.with_param(center_of_rot=center_offset)
 
-    tomogram = axitom.fdk(radiogram, config)
+    tomogram = axitom.fdk(projection, config)
 
     return tomogram
 
@@ -48,8 +46,8 @@ recon_tomo = reconstruct_tomogram()
 correct = axitom.read_image(join(path_to_data, "recon_by_external_software.tif"))
 correct_norm = normalize_grey_scales(correct.transpose())
 
-Reconimg_crop = recon_tomo.transpose()[:, ::-1][1:, :400]
-Reconimg_crop_norm = normalize_grey_scales(Reconimg_crop)
+recon_img_crop = recon_tomo.transpose()[:, ::-1][1:, :400]
+recon_img_crop_norm = normalize_grey_scales(recon_img_crop)
 
 plt.figure()
 plt.subplot(1, 3, 1)
@@ -61,14 +59,14 @@ plt.colorbar()
 
 plt.subplot(1, 3, 2)
 plt.title("AXITOM")
-plt.imshow(Reconimg_crop_norm[::-1, :], cmap=plt.cm.magma)
+plt.imshow(recon_img_crop_norm[::-1, :], cmap=plt.cm.magma)
 plt.axis('off')
 plt.clim(vmin=0.5, vmax=1.0)
 plt.colorbar()
 
 plt.subplot(1, 3, 3)
 plt.title("Absolute deviation")
-plt.imshow(np.abs(Reconimg_crop_norm[::-1, :] - correct_norm), cmap=plt.cm.magma)
+plt.imshow(np.abs(recon_img_crop_norm[::-1, :] - correct_norm), cmap=plt.cm.magma)
 plt.axis('off')
 plt.clim(vmin=0, vmax=0.05)
 plt.colorbar()
