### 2019-09-27 3679306
Simplified the centre of rotation function
