### 2019-09-27 5e79af1
Updated examples
