### 2019-09-28 13ee8f6
Updated tests to new config object
