+++ axitom/tests/test_simulated_projections.py
@@ -13,12 +13,7 @@ class Test_SimulatedProjections(TestCase):
 
         # Set the same settings as used in the forward projection
         param = axitom.config_from_xtekct(dir_path + "/example_data/settings.xtekct")
-        param.n_voxels_x = 500
-        param.n_voxels_y = 500
-        param.n_voxels_z = 500
-        param.n_pixels_u = 500
-        param.n_pixels_v = 500
-        param.update()
+        param = param.with_param(n_pixels_u=500, n_pixels_v=500)
 
         proj = np.load(dir_path + path_to_projections)
         proj = np.exp(-proj)
