### 2019-09-28 80fba16
Updated readme with paths and new config api
