### 2019-09-28 cbe1462
Updated with new paths and new config api
