### 2019-09-28 543a642
Removing legacy code and merging parse into utilities
