+++ axitom/__init__.py
@@ -1,6 +1,5 @@
 from .backprojection import fdk
 from .config import *
-from .parse import *
 from .phantoms import *
 from .projection import forward_project
 from .utilities import *
