### 2019-09-28 e301f0d
Made reconstruction routine "private"
