### 2019-09-28 6c3046b
Updated docs and made some functions "private"
