+++ axitom/__init__.py
@@ -1,5 +1,5 @@
 from .backprojection import fdk
 from .config import *
 from .phantoms import *
-from .projection import forward_project
+from .projection import _forward_project
 from .utilities import *
