+++ axitom/projection.py
@@ -2,41 +2,63 @@ import numpy as np
 from scipy.ndimage import map_coordinates, rotate
 from copy import copy
 
+""" Forward projection routines
 
-def forward_project(data, param, angle_deg):
-
-    # Rotate the data3d around the third axis
-
-    data3d = rotate(data, angle_deg, (1, 0), reshape=False)
-
-    # Now integrate along the rays
-    uu, vv = np.meshgrid(param.detector_us, param.detector_vs)
-
-    Ratios = (param.object_ys + param.source_to_object_dist) / param.source_to_detector_dist
-
-    pus = (uu[:, np.newaxis, :] * Ratios[np.newaxis, :, np.newaxis] - param.object_xs[0]) / param.voxel_size_x
-    pvs = (vv[:, np.newaxis, :] * Ratios[np.newaxis, :, np.newaxis] - param.object_zs[0]) / param.voxel_size_z
-    pys = np.arange(pus.shape[1])[np.newaxis, :, np.newaxis] * np.ones_like(pvs)
-
-    proj2d = np.sum(map_coordinates(data3d[:, :, :], [pvs[:, :, :],
-                                                      pys, pus[:, :, :]], cval=0.,order=3), axis=1)
-
-    dist = np.sqrt(param.source_to_detector_dist ** 2. + uu ** 2. + vv ** 2.) / (
-        param.source_to_detector_dist) * param.voxel_size_y
-
-    return proj2d * dist
+This module contains the functions used to forward project a volume onto a sensor plane. These tools are intended for 
+virtual experiments, validation and testing. 
 
 
-def axis_sym_projection(volume, param, angles=None):
+"""
+
+def _forward_project(volume, config, angle_deg):
     """ Projection of a volume onto a sensor plane using a conical beam geometry
-
-        This implementation has been adapted for axis-symmetry and returns a single projection only
 
         Parameters
         ----------
         volume : np.ndarray
             The volume which will be projected onto the sensor
-        param : obj
+        config : obj
+            The settings object
+        angle_deg : float
+            The angle in degree that the volume will be rotated before the projection is calculated
+
+        Returns
+        -------
+        ndarray
+            The projection
+
+        """
+    data3d = rotate(volume, angle_deg, (1, 0), reshape=False)
+
+    # Now integrate along the rays
+    uu, vv = np.meshgrid(config.detector_us, config.detector_vs)
+
+    Ratios = (config.object_ys + config.source_to_object_dist) / config.source_to_detector_dist
+
+    pus = (uu[:, np.newaxis, :] * Ratios[np.newaxis, :, np.newaxis] - config.object_xs[0]) / config.voxel_size_x
+    pvs = (vv[:, np.newaxis, :] * Ratios[np.newaxis, :, np.newaxis] - config.object_zs[0]) / config.voxel_size_z
+    pys = np.arange(pus.shape[1])[np.newaxis, :, np.newaxis] * np.ones_like(pvs)
+
+    proj2d = np.sum(map_coordinates(data3d[:, :, :], [pvs[:, :, :],
+                                                      pys, pus[:, :, :]], cval=0.,order=3), axis=1)
+
+    dist = np.sqrt(config.source_to_detector_dist ** 2. + uu ** 2. + vv ** 2.) / (
+        config.source_to_detector_dist) * config.voxel_size_y
+
+    return proj2d * dist
+
+
+def axis_sym_projection(volume, config, angles=None):
+    """ Projection of a volume onto a sensor plane using a conical beam geometry
+
+        This implementation has been adapted for axis-symmetry and returns a single projection only.
+        The forward projection is done for several angles and the average projection is returned.
+
+        Parameters
+        ----------
+        volume : np.ndarray
+            The volume which will be projected onto the sensor
+        config : obj
[…]
