### 2019-09-28 e54231d
Cleaned up docstrings
