+++ axitom/backprojection.py
@@ -11,57 +11,50 @@ This module contains the Feldkamp David Kress filtered back projection routines.
 """
 
 
-def map_object_to_detector_coords(object_xs, object_ys, object_zs, settings):
-    """Map the object coordinates to detector pixel coordinates accounting for cone beam divergence.
-
-        This implementation supports both a axis-symmetric mode and full tomograms.
-        This is determined by the shape of the object_xs array.
-
-        The equations and their derivation is found in:
-        Turbell, H. (2001). Cone-Beam Reconstruction Using Filtered Backprojection. Science And Technology.
-        Retrieved from http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.134.5224&amp;rep=rep1&amp;type=pdf
+def map_object_to_detector_coords(object_xs, object_ys, object_zs, config):
+    """ Map the object coordinates to detector pixel coordinates accounting for cone beam divergence.
 
         Parameters
         ----------
         object_xs : np.ndarray
             The x-coordinate array of the object to be reconstructed
         object_ys : np.ndarray
-            The x-coordinate array of the object to be reconstructed
+            The y-coordinate array of the object to be reconstructed
         object_zs : np.ndarray
-            The x-coordinate array of the object to be reconstructed
-        settings : obj
-            The settings object containing all necessary settings for the reconstruction
+            The z-coordinate array of the object to be reconstructed
+        config : obj
+            The config object containing all necessary settings for the reconstruction
 
 
         Returns
         -------
         detector_cords_a
-            The detector coordinates corresponding to the given points
+            The detector coordinates along the a-axis corresponding to the given points
         detector_cords_b
-            The detector coordinates corresponding to the given points
+            The detector coordinates along the b-axis corresponding to the given points
 
         """
 
-    detector_cords_a = (((object_ys * settings.source_to_detector_dist) / (object_xs + settings.source_to_object_dist)) -
-          settings.detector_us[0]) / settings.pixel_size_u
+    detector_cords_a = (((object_ys * config.source_to_detector_dist) / (object_xs + config.source_to_object_dist)) -
+                        config.detector_us[0]) / config.pixel_size_u
 
     if object_xs.ndim == 2:
-        detector_cords_b = (((object_zs[np.newaxis, np.newaxis, :] * settings.source_to_detector_dist) / (
-                object_xs[:, :, np.newaxis] + settings.source_to_object_dist)) - settings.detector_vs[
-                  0]) / settings.pixel_size_v
+        detector_cords_b = (((object_zs[np.newaxis, np.newaxis, :] * config.source_to_detector_dist) / (
+                object_xs[:, :, np.newaxis] + config.source_to_object_dist)) - config.detector_vs[
+                                0]) / config.pixel_size_v
 
     elif object_xs.ndim == 1:
-        detector_cords_b = (((object_zs[np.newaxis, :] * settings.source_to_detector_dist) / (
-                object_xs[:, np.newaxis] + settings.source_to_object_dist)) - settings.detector_vs[
-                  0]) / settings.pixel_size_v
+        detector_cords_b = (((object_zs[np.newaxis, :] * config.source_to_detector_dist) / (
+                object_xs[:, np.newaxis] + config.source_to_object_dist)) - config.detector_vs[
+                                0]) / config.pixel_size_v
     else:
         raise IOError("Invalid dimensions on the object coordinates")
 
     return detector_cords_a, detector_cords_b
 
 
-def _fdk_axisym(projection_filtered, settings):
-    """Filtered back projection algorithm as proposed by Feldkamp David Kress, adapted for axisymmetry.
+def _fdk_axisym(projection_filtered, config):
+    """ Filtered back projection algorithm as proposed by Feldkamp David Kress, adapted for axisymmetry.
 
[…]
