### 2019-09-28 d9c90a8
Cleaning up docstrings
