### 2019-09-28 1930dd2
Cleaning docstrings
