+++ axitom/filtering.py
@@ -29,15 +29,15 @@ def _ramp_kernel_real(cutoff, length):
     return cutoff ** 2.0 * (2.0 * np.sinc(2 * pos * cutoff) - np.sinc(pos * cutoff) ** 2.0)
 
 
-def _add_weights(projection, settings):
+def _add_weights(projection, config):
     """Add weights to the projection according to the ray length traveled through a voxel
 
         Parameters
         ----------
         projection : np.ndarray
             The projection used in the reconstruction
-        settings : obj
-            The settings object containing all necessary settings for the reconstruction
+        config : obj
+            The config object containing all necessary settings for the reconstruction
 
         Returns
         -------
@@ -46,23 +46,23 @@ def _add_weights(projection, settings):
 
 
         """
-    uu, vv = np.meshgrid(settings.detector_vs, settings.detector_us)
+    uu, vv = np.meshgrid(config.detector_vs, config.detector_us)
 
-    weights = settings.source_to_detector_dist / np.sqrt(
-        settings.source_to_detector_dist ** 2. + uu ** 2. + vv ** 2.)
+    weights = config.source_to_detector_dist / np.sqrt(
+        config.source_to_detector_dist ** 2. + uu ** 2. + vv ** 2.)
 
     return projection * weights
 
 
-def ramp_filter_and_weight(projection, settings):
+def ramp_filter_and_weight(projection, config):
     """Add weights to the projection and apply a ramp-high-pass filter set to 0.5*Nyquist_frequency
 
         Parameters
         ----------
         projection : np.ndarray
             The projection used in the reconstruction
-        settings : obj
-            The settings object containing all necessary settings for the reconstruction
+        config : obj
+            The config object containing all necessary settings for the reconstruction
 
         Returns
         -------
@@ -70,7 +70,7 @@ def ramp_filter_and_weight(projection, settings):
             The projections weighted by the ray length and filtered by ramp filter
 
         """
-    projections_weighted = _add_weights(projection, settings)
+    projections_weighted = _add_weights(projection, config)
 
     n_pixels_u, _ = np.shape(projections_weighted)
     ramp_kernel = _ramp_kernel_real(0.5, n_pixels_u)
@@ -82,7 +82,7 @@ def ramp_filter_and_weight(projection, settings):
     for j in range(n_lines):
         projections_filtered[:, j] = sig.fftconvolve(projections_weighted[:, j], ramp_kernel, mode='same')
 
-    scale_factor = 1. / settings.pixel_size_u * np.pi * (
-            settings.source_to_detector_dist / settings.source_to_object_dist)
+    scale_factor = 1. / config.pixel_size_u * np.pi * (
+            config.source_to_detector_dist / config.source_to_object_dist)
 
     return projections_filtered * scale_factor
