### 2019-09-28 b5dcd68
Removed typos
