### 2019-09-28 f5cfbe8
Renaming and removing typos
