+++ axitom/utilities.py
@@ -62,17 +62,17 @@ def _find_center_of_gravity_in_projection(projection, background_internsity=0.9)
         """
     m, n = np.shape(projection)
 
-    binary_radiogram = np.zeros_like(projection, dtype=np.float)
-    binary_radiogram[projection < background_internsity] = 1.
+    binary_proj = np.zeros_like(projection, dtype=np.float)
+    binary_proj[projection < background_internsity] = 1.
 
-    area_x = np.sum(binary_radiogram, axis=1)
-    area_y = np.sum(binary_radiogram, axis=0)
+    area_x = np.sum(binary_proj, axis=1)
+    area_y = np.sum(binary_proj, axis=0)
 
     non_zero_rows = np.arange(n)[area_y != 0.]
     non_zero_columns = np.arange(m)[area_x != 0.]
 
     # Now removing all columns that does not intersect the object
-    object_pixels = binary_radiogram[non_zero_columns, :][:, non_zero_rows]
+    object_pixels = binary_proj[non_zero_columns, :][:, non_zero_rows]
     area_x = area_x[non_zero_columns]
     area_y = area_y[non_zero_rows]
     xs, ys = np.meshgrid(non_zero_rows, non_zero_columns)
@@ -89,7 +89,7 @@ def find_center_of_rotation(projection, background_internsity=0.9, method="cente
         Parameters
         ----------
         projection : ndarray
-            The radiogram, normalized between 0 and 1
+            The projection, normalized between 0 and 1
         background_internsity : float
             The background intensity threshold
         method : string
@@ -105,7 +105,7 @@ def find_center_of_rotation(projection, background_internsity=0.9, method="cente
 
         """
     if projection.ndim != 2:
-        raise ValueError("Invalid radiogram shape. It has to be a 2d numpy array")
+        raise ValueError("Invalid projection shape. It has to be a 2d numpy array")
 
     if method == "center_of_gravity":
         center_v, center_u = _find_center_of_gravity_in_projection(projection, background_internsity)
@@ -159,15 +159,15 @@ def list_files_in_folder(path, file_type=".tif"):
     return natsort.natsorted([file for file in os.listdir(path) if file.endswith(file_type)])
 
 
-def shading_correction(radiograms, flats, darks):
-    """ Perform shading correction on a a list of radiograms based on a list of flat images and list of dark images.
+def shading_correction(projections, flats, darks):
+    """ Perform shading correction on a a list of projections based on a list of flat images and list of dark images.
 
-        The correction is: corrected = (radiogram-dark)/(flat-dark)
+        The correction is: corrected = (projections-dark)/(flat-dark)
 
         Parameters
         ----------
-        radiograms : list
-            A list of radiograms that will be corrected
+        projections : list
+            A list of projections that will be corrected
         flats : list
             A list of flat images
         darks : list
@@ -177,7 +177,7 @@ def shading_correction(radiograms, flats, darks):
         Returns
         -------
         list
-            A list of corrected radiograms
+            A list of corrected projections
 
         """
     flat_avg = np.array(flats)
@@ -186,10 +186,10 @@ def shading_correction(radiograms, flats, darks):
     dark_avg = np.array(darks)
     dark_avg = np.average(dark_avg, axis=0)
 
-    corrected_radiograms = []
-    for radiogram in radiograms:
-        corrected_radiograms.append((radiogram - dark_avg) / (flat_avg - dark_avg))
-    return corrected_radiograms
+    corrected_projections = []
+    for projection in projections:
+        corrected_projections.append((projection - dark_avg) / (flat_avg - dark_avg))
+    return corrected_projections
 
 
 def read_image(file_path, flat_corrected=False):
