### 2019-09-28 db5a6f1
Updated readme format for pypi
