### 2019-09-28 2190572
Centering the logo
