### 2019-09-28 #1704 @PolymerGuy
@dgursoy once again, thank you for the recommended changes. I have now implemented all of them and have in addition done a slight change to the Config class.

I also found a bug which made trouble when the projections did not have square dimensions.

The Config class now contains a smaller set of settings and also has a .with_param() method which returns a copy of the object with the new settings. If I could, I would make the config objects immutable, but as some of the internal parameters are calculated when the object is instantiated, I was not able to do this in a clean way.

The package is now updated on PyPi
