### 2019-09-30 #1704 @katyhuff
@PolymerGuy One DOI (10.1016/j.ultramic.2015.05.002) will need to be fixed in your bibliography. Please remove the 'https://doi.org/' prefix from the DOI itself.
