+++ paper.bib
@@ -68,7 +68,7 @@ volume = "157",
 pages = "35 - 47",
 year = "2015",
 issn = "0304-3991",
-doi = "https://doi.org/10.1016/j.ultramic.2015.05.002",
+doi = "10.1016/j.ultramic.2015.05.002",
 url = "http://www.sciencedirect.com/science/article/pii/S0304399115001060",
 author = "Wim van Aarle and Willem Jan Palenstijn and Jan De Beenhouwer and Thomas Altantzis and Sara Bals and K. Joost Batenburg and Jan Sijbers",
 keywords = "Electron tomography, Reconstruction, ASTRA Toolbox, Dual-axis",
