### 2019-09-30 #1704 @PolymerGuy
@katyhuff I have now removed the prefix from the DOI and pushed the changes to master.
