### 2019-09-30 #1704 @katyhuff
Thank you @PingjunChen and @dgursoy for your reviews -- we couldn't do this without you. Your comments seem to have been particularly valuable to the software package itself. That's wonderful, as it's very much the purpose of JOSS.

Thank you @PolymerGuy for a strong submission and for engaging actively in the review process! I have looked over the paper, double-checked all the DOI links, and have conducted a high-level review of the code itself. Everything now looks ship-shape to me. At this point, please double-check the paper yourself, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. **Please be sure that the DOI metadata (title, authors, etc..) matches this JOSS submission.** Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission. Until then, now is your moment for final touchups! If you bump up the version, please let me know the correct version to associate with this submission.
