### 2019-10-01 #1704 @PolymerGuy
@katyhuff @PingjunChen @dgursoy Thank you all for your kind supervision and helpful advice on my first software submission! I have now archived the repo at Zenodo and got the following DOI:
10.5281/zenodo.3466426

I kept the version as it was during the review.
