### 2019-10-08 #1704 @katyhuff
@openjournals/jose-eics  Hi there, I believe this one is ready for acceptance!
