### 2019-10-08 #1704 @labarba
https://github.com/PolymerGuy/AXITOM/releases
... shows the version as V0.1.3, but here I see V0.1.2 — I will update the version here, but please let us know if that was not what you intended.
