### 2019-10-08 5a37eca
suggested copy editing
https://github.com/openjournals/joss-reviews/issues/1704
