### 2019-10-08 #1704 @labarba
I suggested some copy edits via PR. The text was a bit cluttered and hard to read in places, so I attempted to remove some unnecessary words.
