### 2019-10-09 #1704 @PolymerGuy
@labarba Thank you very much for the PR, and I really appreciate the proposed changes. I have now merged the PR.
