### 2019-10-09 #1704 @labarba
Congratulations, @PolymerGuy, your JOSS paper is now published! 🚀

Huge thanks to our eeditor: @katyhuff, and the reviewers: @PingjunChen, @dgursoy — you're a part of our big adventure!
